+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - The ability to parse GraphQL types and queries from an Add-on's setup file
 - A tag replacement for Structure Nav to better support new template languages and GraphQL
 - A method for embedding native ExpressionEngine templates within Twig and Blade
+- Field content implements countable for better support with Twig's length filter
 
 ### Fixed